+++ tests/invalid/global_variables/multiple_global_defs.rs
@@ -2,10 +2,10 @@
 fail: 1
 */
 
-foo: int = 3;
+foo: i64 = 3;
 
-main(): int {
+main(): i32 {
     return foo;
 }
 
-foo: int = 0;
+foo: i64 = 0;
