+++ tests/invalid/variables/bad_lvalue.rs
@@ -2,8 +2,8 @@
 fail: 1
 */
 
-main(): int {
-    a: int = 2;
+main(): i32 {
+    a: i64 = 2;
     a + 3 = 4;
     return a;
 }
