+++ tests/valid/conditional_expression/ternary_assign.rs
@@ -4,9 +4,9 @@ return: 1
 
 // This is valid, bacause R# uses the C++ ternary operator.
 
-main(): int {
-    a: int = 2;
-    b: int = 1;
+main(): i32 {
+    a: i64 = 2;
+    b: i64 = 1;
     a > b ? a = 1 : a = 0;
     return a;
 }
