+++ tests/valid/conditional_expression/ternary_short_circuit_2.rs
@@ -2,9 +2,9 @@
 return: 2
 */
 
-main(): int {
-    a: int = 0;
-    b: int = 0;
+main(): i32 {
+    a: i64 = 0;
+    b: i64 = 0;
     a ? (b = 1) : (b = 2);
     return b;
 }
