+++ tests/valid/conditional_statement/multiple_if.rs
@@ -2,9 +2,9 @@
 return: 8
 */
 
-main(): int {
-    a: int = 0;
-    b: int = 0;
+main(): i32 {
+    a: i64 = 0;
+    b: i64 = 0;
 
     if (a)
         a = 2;
