+++ tests/valid/functions/expression_args.rs
@@ -2,11 +2,11 @@
 return: 14
 */
 
-add(a: int, b: int): int {
+add(a: i64, b: i64): i64 {
     return a + b;
 }
 
-main(): int {
-    sum: int = add(1 + 2, 4);
+main(): i32 {
+    sum: i64 = add(1 + 2, 4);
     return sum + sum;
 }
