+++ tests/valid/functions/forward_decl_args.rs
@@ -2,12 +2,12 @@
 return: 4
 */
 
-foo(a: int): int;
+foo(a: i64): i64;
 
-main(): int{
+main(): i32{
     return foo(3);
 }
 
-foo(a: int): int{
+foo(a: i64): i64{
     return a + 1;
 }
