+++ tests/valid/functions/variable_as_arg.rs
@@ -2,11 +2,11 @@
 return: 2
 */
 
-foo(x: int): int {
+foo(x: i64): i64 {
     return x + 1;
 }
 
-main(): int {
-    a: int = 1;
+main(): i32 {
+    a: i64 = 1;
     return foo(a);
 }
