+++ tests/valid/loops/break.rs
@@ -2,9 +2,9 @@
 return: 15
 */
 
-main(): int {
-    sum: int = 0;
-    for (i: int = 0; i < 10; i = i + 1) {
+main(): i32 {
+    sum: i64 = 0;
+    for (i: i64 = 0; i < 10; i = i + 1) {
         sum = sum + i;
         if (sum > 10)
             break;
