+++ tests/valid/loops/for_decl.rs
@@ -2,10 +2,10 @@
 return: 3
 */
 
-main(): int {
-    a: int = 0;
+main(): i32 {
+    a: i64 = 0;
 
-    for (i: int = 0; i < 3; i = i + 1)
+    for (i: i64 = 0; i < 3; i = i + 1)
         a = a + 1;
     return a;
 }
