+++ tests/valid/scoping/multi_nesting.rs
@@ -2,11 +2,11 @@
 return: 3
 */
 
-main(): int {
-    a: int = 2;
+main(): i32 {
+    a: i64 = 2;
     if (a < 3) {
         {
-            a: int = 3;
+            a: i64 = 3;
             return a;
         }
         return a;
