+++ tests/valid/variables/exp_return_val.rs
@@ -2,9 +2,9 @@
 return: 0
 */
 
-main(): int {
-    a: int;
-    b: int;
+main(): i32 {
+    a: i64;
+    b: i64;
     a = b = 4;
     return a - b;
 }
